Fixed a 'bug' with setAllBits where I was trying to use memset to set the bits of the bitSet words and not the bytes as memset is supposed to. Changed the testing structure, there are seperate test file for the main functionality of the program then all the tests are ran from a main file. Offer the possibility to pass in arguments when running the tests to set which groups of tests to run.

+++ include/utility.h
@@ -8,8 +8,8 @@ typedef unsigned char bool;
 typedef unsigned char byte;
 
 typedef struct LineClue {
-	byte * clues;
-	byte clueCount;
+	int * clues;
+	int clueCount;
 } LineClue;
 
 LineClue * createLineClueSet (int *, int);
